requirements.txt
pandas==2.1.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
streamlit==1.29.0
xlsxwriter==3.1.9
plotly==5.5.0 